Move QgsRasterLayerProperties class  to the gui API
Motivation :
Move the QgsRasterLayerProperties class to the GUI API.
So, this class could be used using the QGIS API, notably for standalone applications.

The instance of this class was dependent, since QGIS 3.8, from the QgisApp instance, when calling the askUserForDatumTransform(). As this method, defines a title and only call QgsDatumTransformDialog::run() method from the gui API, this PR purpose to :
- move the files of the QgsRasterLayerProperties class ant the QgsMapLayerStyleGuiUtils class to the gui sources
- make those classes public in  the GUI API (GUI_EXPORT)
- change some #include and CMakeList.txt files in accordance
Makes private and documentation

Makes private some methods of the newcomer class in the gui API and fills documentation
Fix SIP and code layout issues




.
.
conflict resolved
.
