+++ python/gui/auto_additions/qgsrasterlayerproperties.py
@@ -1,0 +1,2 @@
+# The following has been generated automatically from src/gui/raster/qgsrasterlayerproperties.h
+QgsRasterLayerProperties.StyleType.baseClass = QgsRasterLayerProperties
