+++ src/gui/qgsmaplayerstyleguiutils.cpp
