+++ src/gui/raster/qgsrasterbandcombobox.h
@@ -18,7 +18,7 @@
 
 #include "qgis_gui.h"
 #include "qgis_sip.h"
-#include "raster/qgsrasterlayer.h"
+#include "qgsrasterlayer.h"
 #include <QComboBox>
 
 class QgsMapLayer;
